+++ Assets/PurrNet/Runtime/BitPacker/Packers/PackUnityTypes.cs
@@ -30,6 +30,20 @@ namespace PurrNet.Packing
         }
 
         [UsedByIL]
+        public static void Write(this BitPacker packer, LayerMask value)
+        {
+            packer.Write((int)value);
+        }
+
+        [UsedByIL]
+        public static void Read(this BitPacker packer, ref LayerMask value)
+        {
+            int val = default;
+            packer.Read(ref val);
+            value = val;
+        }
+
+        [UsedByIL]
         public static void Write(this BitPacker packer, Vector2 value)
         {
             packer.Write(value.x);