+++ Assets/PurrNet/Runtime/CoreModules/Scenes/SceneObjectsModule.cs
@@ -1,15 +1,21 @@
+using System;
 using System.Collections.Generic;
-using UnityEngine;
 using UnityEngine.SceneManagement;
 
 namespace PurrNet.Modules
 {
     public static class SceneObjectsModule
     {
+        public static event Action<Scene> onPreSceneLoad;
+
+        public static event Action<Scene> onPostSceneLoad;
+
         private static readonly List<NetworkIdentity> _sceneIdentities = new List<NetworkIdentity>();
 
         public static void GetSceneIdentities(Scene scene, List<NetworkIdentity> networkIdentities)
         {
+            onPreSceneLoad?.Invoke(scene);
+
             var rootGameObjects = scene.GetRootGameObjects();
 
             PurrSceneInfo sceneInfo = null;
@@ -24,22 +30,13 @@ namespace PurrNet.Modules
             }
 
             if (sceneInfo)
+                rootGameObjects = sceneInfo.rootGameObjects.ToArray();
+
+            for (var i = 0; i < rootGameObjects.Length; i++)
             {
-                var copy = new List<GameObject>(sceneInfo.rootGameObjects);
+                var rootObject = rootGameObjects[i];
 
-                // add any missing root objects
-                foreach (var rootObject in rootGameObjects)
-                {
-                    if (copy.Contains(rootObject)) continue;
-                    copy.Add(rootObject);
-                }
-
-                rootGameObjects = copy.ToArray();
-            }
-
-            foreach (var rootObject in rootGameObjects)
-            {
-                if (rootObject == null || rootObject.scene.handle != scene.handle) continue;
+                if (!rootObject || rootObject.scene.handle != scene.handle) continue;
 
                 rootObject.gameObject.GetComponentsInChildren(true, _sceneIdentities);
 
@@ -48,6 +45,8 @@ namespace PurrNet.Modules
                 rootObject.gameObject.MakeSureAwakeIsCalled();
                 networkIdentities.AddRange(_sceneIdentities);
             }
+
+            onPostSceneLoad?.Invoke(scene);
         }
     }
 }